Make currency amount text watcher a separate class
